add barber CRUD
