+++ UberBarber/UberBarber/Barber/Barber.cs
@@ -11,12 +11,14 @@ namespace UberBarber.Barber
     {
         public Barber(MySqlDataReader reader)
         {
+            BarberId = (int)reader["barber_id"];
             Name = (string)reader["name"];
             Title = (string)reader["title"];
             PhoneNumber = (string)reader["phone_number"];
             Age = (int)reader["age"];
             UserId = (int)reader["user_id"];
         }
+        public int BarberId { get; set; }   
         public string Name { get; set; }
         public string Title { get; set; }
         public string PhoneNumber { get; set; }
